feat(estrutura): Atualiza estrutura backend
